Adjust travis control file.
The `all_branches: true` part is necessary it you want to deploy a tag
from a branch.  Without it, only merges to master are deployed which
means that you cannot apply a tag in a PR and have it deployed.

I also added coverage reporting into the testing command.  This should
enable reporting to codecov.